[OUTLETS] rename Outlets folder
lower case should be used for packages
