+++ app/src/main/java/homer/impl/outlets/COutlet.java
@@ -1,27 +1,27 @@
-package homer.impl.Outlets;
+package homer.impl.outlets;
 
 import homer.api.Device;
 import homer.api.DeviceInfo;
 
 /**
  * Models a standard L-type outlet, which supports a maximum power drain of
- * 3.5kW.
- * See {@link homer.impl.Outlets.Outlet}
+ * 2.0kW.
+ * See {@link homer.impl.outlets.Outlet}
  */
 
-public final class LOutlet extends Outlet {
+public final class COutlet extends Outlet {
 
-    private static final double MAX_VALUE = 3.5;
+    private static final double MAX_VALUE = 2.0;
     private static final double MIN_VALUE = 0;
 
     /**
-     * LOutlet constructor.
+     * COutlet constructor.
      * 
      * @param info
      * @param state
      * @param device
      */
-    public LOutlet(final DeviceInfo info, final double state, final Device<?> device) {
+    public COutlet(final DeviceInfo info, final double state, final Device<?> device) {
         super(info, state, MIN_VALUE, MAX_VALUE, device);
     }
